IG-3-2: вынесен EmojiBar в отдельный компонент
