+++ src/common/components/EmojiBar/index.ts
@@ -1,0 +1,1 @@
+export { EmojiBar } from "./EmojiBar";
