+++ src/features/Messages/component/ChatSection/InputBar/InputBar.tsx
@@ -1,16 +1,14 @@
 import React, { useState } from "react";
 import { Box, IconButton, Stack } from "@mui/material";
 import styled from "./InputBar.module.scss";
-import Picker, { IEmojiData } from "emoji-picker-react";
-import { ReactComponent as Smile } from "../../../../../assets/svg/smile.svg";
 import { ReactComponent as UploadImages } from "../../../../../assets/svg/uploadImages.svg";
+import { EmojiBar } from "../../../../../common/components/EmojiBar";
 
 interface IProps {
   sendMessage: (inputStr: string) => void;
 }
 
 export const InputBar: React.FC<IProps> = ({ sendMessage }) => {
-  const [showEmoji, setShowEmoji] = useState(false);
   const [inputStr, setInputStr] = useState("");
 
   const handleChangeText = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -26,11 +24,6 @@ export const InputBar: React.FC<IProps> = ({ sendMessage }) => {
     }
   };
 
-  const onEmojiClick = (event: React.MouseEvent, emojiObject: IEmojiData) => {
-    setInputStr((prevInput) => prevInput + emojiObject.emoji);
-    setShowEmoji(false);
-  };
-
   return (
     <Stack className={styled.inputBar}>
       <Box component="form" onSubmit={handleSendMessage}>
@@ -40,15 +33,7 @@ export const InputBar: React.FC<IProps> = ({ sendMessage }) => {
           alignItems="center"
           flexDirection="row"
         >
-          <div className={styled.emojiPicker}>
-            {showEmoji && <Picker onEmojiClick={onEmojiClick} />}
-          </div>
-          <button
-            className={styled.btnEmoji}
-            onClick={() => setShowEmoji((prevState) => !prevState)}
-          >
-            <Smile />
-          </button>
+          <EmojiBar setInputStr={setInputStr} />
           <input
             placeholder="Message..."
             className={styled.inputText}
