+++ src/features/Messages/component/ChatSection/index.ts
@@ -1,0 +1,1 @@
+export { ChatSection } from "./ChatSection";
